+++ src/platform.ts
@@ -58,7 +58,6 @@ export class SmartHQPlatform implements DynamicPlatformPlugin {
   platformLogging!: options['logging']
   platformRefreshRate!: options['refreshRate']
   platformPushRate!: options['pushRate']
-  platformUpdateRate!: options['updateRate']
   version!: string
 
   // Matter support tracking
@@ -1547,10 +1546,6 @@ export class SmartHQPlatform implements DynamicPlatformPlugin {
     this.platformRefreshRate = this.config.options?.refreshRate ? this.config.options.refreshRate : undefined
     const refreshRate = this.config.options?.refreshRate ? 'Using Platform Config refreshRate' : 'Platform Config refreshRate Not Set'
     await this.debugLog(`${refreshRate}: ${this.platformRefreshRate}`)
-    // UpdateRate
-    this.platformUpdateRate = this.config.options?.updateRate ? (this.config.options.updateRate * 1000) : undefined
-    const updateRateMsg = this.config.options?.updateRate ? 'Using Platform Config updateRate' : 'Platform Config updateRate Not Set'
-    await this.debugLog(`${updateRateMsg}: ${this.platformUpdateRate}`)
     // PushRate
     this.platformPushRate = this.config.options?.pushRate ? this.config.options.pushRate : undefined
     const pushRate = this.config.options?.pushRate ? 'Using Platform Config pushRate' : 'Platform Config pushRate Not Set'